CloudComputing/requirements.txt
flask==2.3.2
tensorflow==2.18.0
numpy>=1.25.0
Pillow==9.4.0

gunicorn==20.1.0
google-cloud-storage==2.9.0
